+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.2 (Build Yet...)
 * Modify the wrong word from README. 'Isnow.strandard' ->  'IsNow.standard'
+* The build result in the browser is umd.
 
 ## 1.0.1
 * Modify the wrong word from README.